requirements.txt
nose
sphinx
falcon
gunicorn
httpie
uvicorn
aiohttp
pytest
pytest-asyncio
pytest-mock
pep8
requests
termcolor
python-crontab
schedule
solana
python-dotenv